+++ app/constants/IMLocalize.js
@@ -1,0 +1,56 @@
+import i18n from 'i18next';
+import {initReactI18next} from 'react-i18next';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import * as RNLocalize from 'react-native-localize';
+
+import en from './translations/en';
+import es from './translations/es';
+
+const LANGUAGES = {
+  en,
+  es,
+};
+
+const LANG_CODES = Object.keys(LANGUAGES);
+
+const LANGUAGE_DETECTOR = {
+  type: 'languageDetector',
+  async: true,
+  detect: callback => {
+    AsyncStorage.getItem('user-language', (err, language) => {
+      // if error fetching stored data or no language was stored
+      // display errors when in DEV mode as console statements
+      if (err || !language) {
+        if (err) {
+          console.log('Error fetching Languages from asyncstorage ', err);
+        } else {
+          console.log('No language is set, choosing English as fallback');
+        }
+        const findBestAvailableLanguage =
+          RNLocalize.findBestAvailableLanguage(LANG_CODES);
+
+        callback(findBestAvailableLanguage.languageTag || 'en');
+        return;
+      }
+      callback(language);
+    });
+  },
+  init: () => {},
+  cacheUserLanguage: language => {
+    AsyncStorage.setItem('user-language', language);
+  },
+};
+
+i18n
+  .use(LANGUAGE_DETECTOR)
+  .use(initReactI18next)
+  .init({
+    compatibilityJSON: 'v3',
+    resources: LANGUAGES,
+    react: {
+      useSuspense: false,
+    },
+    interpolation: {
+      escapeValue: false,
+    },
+  });