+++ app/constants/translations/en/navigate.js
@@ -1,0 +1,11 @@
+export default {
+    home: {
+        title: 'Home',
+    },
+    book: {
+        title: 'Book',
+    },
+    setting: {
+        title: 'Setting',
+    },
+};