+++ app/constants/translations/es/common.js
@@ -1,0 +1,4 @@
+export default {
+  hello: 'Hola',
+  languageSelector: 'Elige tu idioma',
+};