+++ app/constants/translations/es/navigate.js
@@ -1,0 +1,11 @@
+export default {
+    home: {
+        title: 'Inicio',
+    },
+    book: {
+        title: 'Libro',
+    },
+    setting: {
+        title: 'Configuración',
+    },
+};